chore: Make Kiiroo emulator use controls device panel
Use Device panel from controls library instead of own implementation
