+++ Buttplug.Apps.KiirooEmulatorGUI/KiirooEmulatorPanel.xaml
@@ -7,11 +7,10 @@
              d:DesignHeight="300" d:DesignWidth="300">
     <Grid Background="#FFE5E5E5">
         <Grid.RowDefinitions>
-            <RowDefinition Height="*" />
             <RowDefinition Height="20" />
+            <RowDefinition Height="20" />
+            <RowDefinition Height="*" />
         </Grid.RowDefinitions>
-        <ListBox Grid.Row="0" Name="DeviceListBox" Margin="10,10,10,10" />
-        <Button Grid.Row="1" Name="ScanButton" Content="Start Scanning" Margin="10,0,0,0" HorizontalAlignment="Left" Width="95" Click="ScanButton_Click" />
-        <Button Grid.Row="1" Name="ServerButton" Content="Start Server" Margin="0,0,10,0" HorizontalAlignment="Right" Width="95" Click="ServerButton_Click" />
+        <Button Grid.Row="1" Name="ServerButton" Content="Stop Server" Margin="0,0,195,0" HorizontalAlignment="Right" Width="95" Click="ServerButton_Click" />
     </Grid>
 </UserControl>
